+++ src/convtools/contrib/models/utils.py
@@ -1,26 +1,32 @@
 """Defines methods which should close the gap in typing functionality between
 python 3.6 and 3.7+ """
 # flake8: noqa
+# pylint: disable=ungrouped-imports
 import sys
-from threading import RLock
-from typing import Dict, List, Tuple, Union
+from typing import TYPE_CHECKING
 
 from .base import BaseModel, CastOverrides
 
 
+if TYPE_CHECKING:
+    from typing import Callable, Optional
+
+
 PY_VERSION = sys.version_info[0:2]
 
-UnionCls = Union[int, str].__class__
 
 if PY_VERSION <= (3, 6):
 
+    from typing import Dict, List, Set, Tuple, Union
+
+    UnionCls = Union[int, str].__class__
     GenericAlias = type(List[int])
 
-    def is_generic_alias(obj):
-        return isinstance(obj, GenericAlias) or is_generic_union_alias(obj)
-
-    def is_generic_union_alias(obj):
+    def is_union(obj):
         return obj.__class__ is UnionCls
+
+    def is_generic_alias(obj):
+        return isinstance(obj, GenericAlias)
 
     def is_generic_list_alias(alias):
         return alias.__origin__ is List
@@ -31,14 +37,24 @@ if PY_VERSION <= (3, 6):
     def is_generic_dict_alias(alias):
         return alias.__origin__ is Dict
 
-else:  # if PY_VERSION > (3, 6): -- comment for .coveragerc
-    from typing import _GenericAlias  # type: ignore
+    def is_generic_set_alias(alias):
+        return alias.__origin__ is Set
+
+elif (
+    (3, 7)
+    <= PY_VERSION
+    <= (
+        3,
+        8,
+    )
+):  # if PY_VERSION > (3, 6): -- comment for .coveragerc
+    from typing import Union, _GenericAlias  # type: ignore
+
+    def is_union(obj):
+        return isinstance(obj, _GenericAlias) and obj.__origin__ is Union
 
     def is_generic_alias(obj):
         return isinstance(obj, _GenericAlias)
-
-    def is_generic_union_alias(obj):
-        return obj.__origin__ is Union
 
     def is_generic_list_alias(alias):
         return alias.__origin__ is list
@@ -48,6 +64,69 @@ else:  # if PY_VERSION > (3, 6): -- comment for .coveragerc
 
     def is_generic_dict_alias(alias):
         return alias.__origin__ is dict
+
+    def is_generic_set_alias(alias):
+        return alias.__origin__ is set
+
+elif PY_VERSION == (3, 9):
+    from typing import (  # type: ignore
+        Dict,
+        GenericAlias,
+        List,
+        Tuple,
+        Union,
+        _GenericAlias,
+        _UnionGenericAlias,
+    )
+
+    alias_types = (GenericAlias, _GenericAlias)
+
+    def is_union(obj):
+        return isinstance(obj, _UnionGenericAlias)
+
+    def is_generic_alias(obj):
+        return isinstance(obj, alias_types)
+
+    def is_generic_list_alias(alias):
+        return alias.__origin__ is list
+
+    def is_generic_tuple_alias(alias):
+        return alias.__origin__ is tuple
+
+    def is_generic_dict_alias(alias):
+        return alias.__origin__ is dict
+
+    def is_generic_set_alias(alias):
+        return alias.__origin__ is set
+
+elif PY_VERSION >= (3, 10):
+    from types import UnionType  # type: ignore
+    from typing import (  # type: ignore
+        GenericAlias,
+        _GenericAlias,
+        _UnionGenericAlias,
+    )
+
+    alias_types = (GenericAlias, _GenericAlias)
+    union_types = (UnionType, _UnionGenericAlias)
+
+    def is_union(obj):
+        return isinstance(obj, union_types)
+
+    def is_generic_alias(obj):
+        return isinstance(obj, alias_types)
+
+    def is_generic_list_alias(alias):
+        return alias.__origin__ is list
+
+    def is_generic_tuple_alias(alias):
+        return alias.__origin__ is tuple
+
+    def is_generic_dict_alias(alias):
+        return alias.__origin__ is dict
+
+    def is_generic_set_alias(alias):
+        return alias.__origin__ is set
 
 
 if PY_VERSION <= (3, 7):
@@ -79,23 +158,17 @@ def is_model_cls(cls):
     return isinstance(cls, type) and issubclass(cls, BaseModel)
 
 
-def __hash__(self):  # pylint: disable=invalid-name
-    return hash(tuple(self.__args__))
-
-
-patching_lock = RLock()
-
-
 class TypeValueWrapper:
     """We have to patch typing.Union hash method to make it order sensitive as
     it is required for lru_cache + Union type casting"""
 
-    __slots__ = ("type_value", "cast", "cast_overrides")
+    __slots__ = ("type_value", "cast", "cast_overrides", "union_args_getter")
 
     def __init__(self, type_value, cast: bool, cast_overrides: CastOverrides):
         self.type_value = type_value
         self.cast = cast
         self.cast_overrides = cast_overrides
+        self.union_args_getter = None
 
     def validate_args(self):
         if not self.cast and self.cast_overrides:
@@ -111,13 +184,18 @@ class TypeValueWrapper:
             self.type_value == other.type_value
             and self.cast == other.cast
             and self.cast_overrides == other.cast_overrides
+            and (
+                self.union_args_getter is None
+                or (
+                    self.union_args_getter(  # pylint: disable=not-callable
+                        self.type_value
+                    )
+                    == self.union_args_getter(  # pylint: disable=not-callable
+                        other.type_value
+                    )
+                )
+            )
         )
 
     def __hash__(self):
-        with patching_lock:
-            prev_hash = UnionCls.__hash__
-            UnionCls.__hash__ = __hash__
-            try:
-                return hash(self.type_value)
-            finally:
-                UnionCls.__hash__ = prev_hash
+        return hash(self.type_value)